Animated Ghost
An animated ghost I recoded. Found it on instagram, shared by @webcrafters.cz. Took a while to make it work because of a rather stupid mistake I made and couldn't find in the first place. 
Next step would be to refacture the code using classes instead of id's. 
Code review by @kidzki <3.